pyproject.toml
[project]
name = "repo"
version = "0.1.0"
description = "Autopoietic Meta-Agent Orchestrator System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.20",
    "typing-extensions>=4.9.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
